FeatureHub toggles for Measurement

+++ MeasurementService/Controllers/MeasurementController.cs
@@ -1,5 +1,6 @@
 ﻿using DefaultNamespace;
 using Microsoft.AspNetCore.Mvc;
+using Globalization;
 
 namespace MeasurementService.Controllers
 {
@@ -25,6 +26,12 @@ namespace MeasurementService.Controllers
         [HttpGet("GetAll/{ssn}")]
         public async Task<IActionResult> GetAllMeasurements(string ssn)
         {
+            var enabled = await getDeploymentToggle();
+            if(!enabled)
+            {
+                return BadRequest("Feature is disabled");
+            }
+
             var measurements = await _measurementService.GetAllMeasurement(ssn);
             return Ok(measurements);
         }
@@ -32,6 +39,12 @@ namespace MeasurementService.Controllers
         [HttpGet("GetById/{id}")]
         public async Task<IActionResult> GetMeasurementById(int id)
         {
+            var enabled = await getDeploymentToggle();
+            if(!enabled)
+            {
+                return BadRequest("Feature is disabled");
+            }
+
             var measurement = await _measurementService.GetMeasurementById(id);
             if (measurement == null)
                 return NotFound($"No measurement found with ID {id}.");
@@ -40,17 +53,16 @@ namespace MeasurementService.Controllers
         }
 
         [HttpPost("Create")]
-        public async Task<IActionResult> CreateMeasurement(MeasurementDto measurementDto)
+        public async Task<IActionResult> CreateMeasurement(MeasurementDto measurementDto, [FromHeader(Name = "country")] string country)
         {
-           
+
+            var countryName = CountryConverter.GetCountryName(country);
+
             var config = await _featureHubContext.GetFeatureHubContextAsync();
 
+            var context = await config.NewContext().Attr("country", countryName.ToLower()).Build();
             
-            
-            var context = await config.NewContext().Attr("country", "denmark").Build();
-            
-     
-            var enbabled =  context["CreateMeasurement"].IsEnabled;
+            var enbabled = context["CreateMeasurement"].IsEnabled;
             if(!enbabled)
             {
                 return BadRequest("Feature is disabled");
@@ -66,6 +78,13 @@ namespace MeasurementService.Controllers
         [HttpPut("Update/{id}")]
         public async Task<IActionResult> UpdateMeasurement(MeasurementDto measurementDto, int id)
         {
+            
+            var enabled = await getDeploymentToggle();
+            if(!enabled)
+            {
+                return BadRequest("Feature is disabled");
+            }
+
             var updatedMeasurement = await _measurementService.UpdateMeasurement(
                 measurementDto,
                 id
@@ -79,6 +98,11 @@ namespace MeasurementService.Controllers
         [HttpDelete("Delete/{id}")]
         public async Task<IActionResult> DeleteMeasurement(int id)
         {
+            var enabled = await getDeploymentToggle();
+            if(!enabled)
+            {
+                return BadRequest("Feature is disabled");
+            }
             await _measurementService.DeleteMeasurement(id);
             return NoContent();
         }
@@ -86,8 +110,22 @@ namespace MeasurementService.Controllers
         [HttpPost("RebuildDb")]
         public IActionResult RebuildDb()
         {
+            var enabled = getDeploymentToggle().Result;
+            if(!enabled)
+            {
+                return BadRequest("Feature is disabled");
+            }
             _measurementService.RebuildDb();
             return Ok("Database rebuilt successfully.");
+        }
+
+        private async Task<bool> getDeploymentToggle()
+        {
+            var config = await _featureHubContext.GetFeatureHubContextAsync();
+
+            var context = await config.NewContext().Build();
+
+            return context["Production"].IsEnabled;
         }
     }
 }
